+++ nsd/connio.c
@@ -165,7 +165,7 @@ Ns_ConnFlush(Ns_Conn *conn, char *buf, int len, int stream)
     if (!stream
 	    && (conn->flags & NS_CONN_GZIP)
 	    && (servPtr->opts.flags & SERV_GZIP)
-	    && (len > servPtr->opts.gzipmin)
+	    && (len > (int) servPtr->opts.gzipmin)
 	    && (ahdr = Ns_SetIGet(conn->headers, "Accept-Encoding")) != NULL
 	    && strstr(ahdr, "gzip") != NULL
 	    && Ns_Compress(buf, len, &gzip, servPtr->opts.gziplevel) == NS_OK) {
